bangla-plagiarism-checker/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# PDF Processing
PyMuPDF==1.23.8
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78

# Text Processing & NLP
nltk==3.8.1
spacy==3.7.2
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
numpy==1.25.2
pandas==2.1.4

# Utilities
requests==2.31.0
aiofiles==23.2.0
httpx==0.25.2
celery==5.3.4